+++ src/main/java/com/chatop/api/repository/RentalRepository.java
@@ -3,5 +3,5 @@ package com.chatop.api.repository;
 import com.chatop.api.entity.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
 }
